- Removed ushort from SchematronStore - Better exception handling in CrlLookup
git-svn-id: https://svn.softwareboersen.dk/openebusiness/dk.gov.oiosi/library/dotnet/trunk@17059 4e58de57-8926-0410-947e-8945c843cdd7

+++ src/dk.gov.oiosi/xml/schematron/ISchematronStoreConfig.cs
@@ -7,6 +7,6 @@ namespace dk.gov.oiosi.xml.schematron {
         /// <summary>
         /// Gets the maximum number of compiled stylesheet to have in the memory
         /// </summary>
-        ushort MaxCompiledStylesheetsInMemory { get; }
+        int MaxCompiledStylesheetsInMemory { get; }
     }
 }
