+++ src/dk.gov.oiosi/xml/schematron/SchematronStoreConfig.cs
@@ -7,14 +7,14 @@ namespace dk.gov.oiosi.xml.schematron {
     /// </summary>
     [XmlRoot(Namespace = ConfigurationHandler.RaspNamespaceUrl)]
     public class SchematronStoreConfig : ISchematronStoreConfig {
-        private ushort _maxCompiledStylesheetsInMemory = 2;
+        private int _maxCompiledStylesheetsInMemory = 2;
 
         #region ISchematronStoreConfig Members
 
         /// <summary>
         /// Gets and sets the max compiled stylesheets to have in memory.
         /// </summary>
-        public ushort MaxCompiledStylesheetsInMemory {
+        public int MaxCompiledStylesheetsInMemory {
             get { return _maxCompiledStylesheetsInMemory; }
             set { _maxCompiledStylesheetsInMemory = value; }
         }
